requirements.txt
pydantic
sqlalchemy
fastapi
python-jose
uliweb-alembic
chromadb
langchain-openai
langchain-community
langchain-text-splitters
langchain-chroma